TASK-3_2: Cannot build scaffold as needed
